+++ integration-tests/public/custom-environment/lib.rs
@@ -14,6 +14,10 @@ pub enum EnvironmentWithManyTopics {}
 impl Environment for EnvironmentWithManyTopics {
     const NATIVE_TO_ETH_RATIO: u32 =
         <DefaultEnvironment as Environment>::NATIVE_TO_ETH_RATIO;
+    const TRUST_BACKED_ASSETS_PRECOMPILE_INDEX: u16 =
+        <DefaultEnvironment as Environment>::TRUST_BACKED_ASSETS_PRECOMPILE_INDEX;
+    const POOL_ASSETS_PRECOMPILE_INDEX: u16 =
+        <DefaultEnvironment as Environment>::POOL_ASSETS_PRECOMPILE_INDEX;
 
     type AccountId = <DefaultEnvironment as Environment>::AccountId;
     type Balance = <DefaultEnvironment as Environment>::Balance;